+++ packages/theme/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Patch Changes
 
 - chore: update dependencies
-- fix: sass warn
+- fix: sass warn - [legacy-js-api](https://sass-lang.com/documentation/breaking-changes/legacy-js-api/)
 
 ## 0.5.1